+++ src/guppy_control/src/t200_interface.cpp
@@ -19,12 +19,12 @@ bool T200Interface::setup_can() {
     return true;
 }
 
-bool T200Interface::send_to_can(unsigned int can_id, double value) {
+bool T200Interface::send_to_can(unsigned int can_id, float value) {
     struct can_frame frame;
     frame.can_id = can_id;
-    frame.can_dlc = sizeof(double);
+    frame.can_dlc = sizeof(float);
 
-    std::memcpy(frame.data, &value, sizeof(double));
+    std::memcpy(frame.data, &value, sizeof(float));
 
     if (::write(sock_, &frame, sizeof(struct can_frame)) < 0) return false;
     return true;
@@ -36,8 +36,8 @@ bool T200Interface::write(Eigen::VectorXd throttles) {
 
     bool okay = true;
     for (size_t i=0; i < n; i++) {
-        okay = okay && send_to_can(can_ids[i], throttles[i]);
-        // std::cout << can_ids[i] << "\t" << throttles[i] << "\t" << okay << std::endl;
+        okay = okay && send_to_can(can_ids[i], (float)throttles[i]);
+        // std::cout << can_ids[i] << "\t" << (float)throttles[i] << "\t" << okay << std::endl;
     }
 
     return okay;
